MAJOR CHANGES to filter.bpfilt/lpfilt/hpfilt
The original implementation has been replaced with a much simpler one,
which may be less accurate in the design of the filters, but that is
more robust. The change was motivated by the realization that the
original implementation of hpfilt could introduce very large
artifacts in the data in the presence of sharp signal changes.