New node for fast decimation of physiosets
Node decimate can be used to reduce the sampling rate of the
input physioset. Note that the decimate node does not perform
any anti-aliasing filtering. It is diffult to assess how much
faster the decimate node will be with respect to the more
general resample node. I would expect least be at least 3 or 4
times faster.